chore(links): refactored vote buttons into component
