+++ client/components/voteButtons.js
@@ -1,0 +1,62 @@
+import React from 'react';
+import { Link } from 'react-router';
+import { connect } from 'react-redux';
+import { updateVote } from '../reducer/link';
+
+const VoteButtons = (props) =>
+  (
+    <div>
+      {
+        props.loggedIn
+          ? <div>
+              <div className="row">
+                <div className="col-md-6">
+                  <button className="upvote btn-success"  onClick={() => props.dispatchUpvote(props.item)}>Upvote</button>
+                </div>
+              </div>
+
+              <div className="row">
+                <div className="col-md-6">
+                  <button className="downvote btn-success"  onClick={() => props.dispatchDownvote(props.item)}>Downvote</button>
+                </div>
+              </div>
+            </div>
+          : <div>
+              <h4>
+                <Link className="loginSignupButton" to='/login'><strong>Log in </strong></Link>or
+                <Link className="loginSignupButton" to='/signup'><strong> sign up </strong></Link> to vote!
+              </h4>
+            </div>
+      }
+    </div>
+  );
+
+
+const mapDispatch = (dispatch) => ({
+  dispatchUpvote: link => {
+    const info = {
+      error: link.error,
+      vendor: link.vendor,
+      vendor_id: link.vendor_id,
+      vote: 'upvote',
+      url: link.url,
+      created: link.created,
+      modified: link.modified
+    }
+    dispatch(updateVote(info));
+  },
+  dispatchDownvote: link => {
+    const info = {
+      error: link.error,
+      vendor: link.vendor,
+      vendor_id: link.vendor_id,
+      vote: 'downvote',
+      url: link.url,
+      created: link.created,
+      modified: link.modified
+    }
+    dispatch(updateVote(info));
+  }
+});
+
+export default connect(null, mapDispatch)(VoteButtons);
